+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.9.7] - 2020-06-12
+### Added
+- VHD Control multi VHDs mount (still needs some work).
 ### Changed
 - Directory structure for VBIOS extraction, iommu and VM scripts.
 ### Fixed
@@ -12,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Custom VM ISO selection.
 - Display check logic (checks only if multi GPU configuration is detected).
 - GPU method handling with custom names.
+- VHD Control.
 
 ## [0.9.6] - 2020-06-11
 ### Added